requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-PyMongo==2.3.0
pymongo==3.11.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.0.0
pytz==2020.5
requests==2.25.1
six==1.15.0
twilio==6.50.1
urllib3==1.26.2
Werkzeug==1.0.1
gunicorn
flask_login
python-dotenv
currencyconverter
pytest
Flask-MongoAlchemy==0.7.1
google-cloud
google-cloud-storage